minimax, heuristic, findChosenMove functions
Implements the minimax algorithm with alpha-beta pruning, the heuristic evaluator helper function, as well as findChosenMove to determine which next legal move is to be chosen. NOTE: the current heuristic function only calculates a score based on the game score and corners. Upon better understanding of what getStability returns will it be added to the calculation.